requirements.txt
asgiref==3.4.1
bleach==3.3.1
certifi==2021.5.30
charset-normalizer==2.0.3
colorama==0.4.4
Django==3.2.5
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-translations==1.3.0
djangorestframework==3.12.4
drf-api-logger==1.0.8
gunicorn==20.1.0
idna==3.2
packaging==21.0
Pillow==8.3.1
psycopg2==2.9.1
pyparsing==2.4.7
pyTelegramBotAPI==3.8.2
python-dotenv==0.18.0
pytz==2021.1
qrcode==7.2
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
webencodings==0.5.1
